client redemption api update
