+++ src/client_redemption/client_channel_managers/client_channel_remoteapp_manager.hpp
@@ -29,6 +29,7 @@
 #include "core/RDP/orders/AlternateSecondaryWindowing.hpp"
 
 #include "client_redemption/client_redemption_api.hpp"
+#include "client_redemption/client_input_output_api/client_graphic_api.hpp"
 
 
 // [MS-RDPERP]: Remote Desktop Protocol: Remote Programs Virtual Channel Extension
